+++ assistants/core/src/index.ts
@@ -2,6 +2,7 @@ import {
   AssistantPackage,
   RuleDefinition,
   AssistantDefinition,
+  AssistantEnv,
 } from '@sketch-hq/sketch-assistant-types'
 import { I18n, setupI18n } from '@lingui/core'
 
@@ -25,8 +26,8 @@ import * as layersNoHidden from './rules/layers-no-hidden'
 import * as layersNoLoose from './rules/layers-no-loose'
 import * as layersSubpixelPositioning from './rules/layers-subpixel-positioning'
 import * as layerStylesNoDirty from './rules/layer-styles-no-dirty'
-import * as layerStylesPreferShared from './rules/layer-styles-prefer-shared'
 import * as layerStylesPreferLibrary from './rules/layer-styles-prefer-library'
+import * as layerStylesPreferShared from './rules/layer-styles-prefer-shared'
 import * as namePatternArtboards from './rules/name-pattern-artboards'
 import * as namePatternGroups from './rules/name-pattern-groups'
 import * as namePatternImages from './rules/name-pattern-images'
@@ -37,6 +38,7 @@ import * as namePatternText from './rules/name-pattern-text'
 import * as resultMessagesInclude from './rules/result-messages-include'
 import * as shadowsNoDisabled from './rules/shadows-no-disabled'
 import * as sharedStylesNoUnused from './rules/shared-styles-no-unused'
+import * as symbolsNoDetached from './rules/symbols-no-detached'
 import * as symbolsNoUnused from './rules/symbols-no-unused'
 import * as symbolsPreferLibrary from './rules/symbols-prefer-library'
 import * as textStylesNoDirty from './rules/text-styles-no-dirty'
@@ -48,18 +50,21 @@ import zhHansMessages from './locale/zh-Hans/messages'
 
 export type CreateRuleFunction = (i18n: I18n) => RuleDefinition
 
-const SUPPORTED_LOCALES = ['en', 'zh-Hans']
-const FALLBACK_LOCALE = 'en'
-const pkgName = '@sketch-hq/sketch-core-assistant'
-
-const assistant: AssistantPackage = async (env) => {
-  const i18n: I18n = setupI18n({
+export const createI18NObject = (env: AssistantEnv): I18n => {
+  const SUPPORTED_LOCALES = ['en', 'zh-Hans']
+  const FALLBACK_LOCALE = 'en'
+  return setupI18n({
     language: SUPPORTED_LOCALES.includes(env.locale!) ? env.locale : FALLBACK_LOCALE,
     catalogs: {
       en: enMessages,
       'zh-Hans': zhHansMessages,
     },
   })
+}
+
+const assistant: AssistantPackage = async (env) => {
+  const i18n: I18n = createI18NObject(env)
+  const pkgName = '@sketch-hq/sketch-core-assistant'
   const definition: AssistantDefinition = {
     name: pkgName,
     rules: [
@@ -95,6 +100,7 @@ const assistant: AssistantPackage = async (env) => {
       resultMessagesInclude,
       shadowsNoDisabled,
       sharedStylesNoUnused,
+      symbolsNoDetached,
       symbolsNoUnused,
       symbolsPreferLibrary,
       textStylesNoDirty,