+++ packages/utils/src/files/index.ts
@@ -34,7 +34,6 @@ const fromFile = async (filepath: string): Promise<SketchFile> => {
         },
         meta: JSON.parse(archive.entryDataSync('meta.json').toString()),
         user: JSON.parse(archive.entryDataSync('user.json').toString()),
-        // @ts-ignore TODO: Fix this, `assistants` prop should be optional in the file format
         workspace,
       })
     })
@@ -45,4 +44,20 @@ const fromFile = async (filepath: string): Promise<SketchFile> => {
   return { filepath, contents }
 }
 
-export { fromFile }
+/**
+ * Filter pages out of a SketchFile object based on page ids.
+ */
+const filterPages = (file: SketchFile, excludedPageIds: string[]): SketchFile => ({
+  ...file,
+  contents: {
+    ...file.contents,
+    document: {
+      ...file.contents.document,
+      pages: file.contents.document.pages.filter(
+        (page) => !excludedPageIds.includes(page.do_objectID),
+      ),
+    },
+  },
+})
+
+export { fromFile, filterPages }