+++ packages/utils/src/run/index.ts
@@ -1,2 +1,4 @@
 export * from './run-assistant'
 export * from './run-multiple-assistants'
+export * from './profile'
+export * from './ignore'