+++ assistants/core/CHANGELOG.md
@@ -1,5 +1,95 @@
 # @sketch-hq/sketch-core-assistant
 
+## 5.0.0
+
+### Major Changes
+
+- 6733a97: Numerous refactors around Sketch file object iteration.
+- e76fabd: Introduction of the Assistants CLI.
+
+### Minor Changes
+
+- 76c8ded: Add the `symbols-no-detached` rule.
+
+### Patch Changes
+
+- c4615bc: Update rule copy
+- 1698a52: Fix rule title copy in `text-styles-prefer-shared`.
+- 1c388d9: Add core rules to monorepo.
+- 7a54cbb: Count symbols set as override values as used in `symbols-no-unused`.
+- 04b71a6: Fix `groups-no-empty` flagging an empty document.
+- caba0e1: Do not raise violations from the `exported-layers-normal-blend-mode` rule when the layer
+  is only to be exported as some kind of bitmap-based export format.
+- 0688585: Report loose layers, not the parent page in the the `layers-no-loose` rule.
+- 581554d: Fix a bug where violation objects weren't passed correctly.
+- Updated dependencies [8351103]
+- Updated dependencies [67f62f2]
+- Updated dependencies [c802b6b]
+- Updated dependencies [21de31c]
+- Updated dependencies [a004097]
+- Updated dependencies [e76fabd]
+- Updated dependencies [2f6f82b]
+- Updated dependencies [6733a97]
+- Updated dependencies [bc13811]
+  - @sketch-hq/sketch-assistant-types@5.0.0
+
+## 5.0.0-next.22
+
+## 5.0.0-next.21
+
+### Patch Changes
+
+- Updated dependencies [21de31c]
+  - @sketch-hq/sketch-assistant-types@5.0.0-next.21
+
+## 5.0.0-next.20
+
+## 5.0.0-next.19
+
+### Patch Changes
+
+- 04b71a6: Fix `groups-no-empty` flagging an empty document.
+- Updated dependencies [c802b6b]
+- Updated dependencies [a004097]
+  - @sketch-hq/sketch-assistant-types@5.0.0-next.19
+
+## 5.0.0-next.18
+
+### Patch Changes
+
+- Updated dependencies [2f6f82b]
+  - @sketch-hq/sketch-assistant-types@5.0.0-next.18
+
+## 5.0.0-next.17
+
+### Patch Changes
+
+- c4615bc: Update rule copy
+- Updated dependencies [67f62f2]
+  - @sketch-hq/sketch-assistant-types@5.0.0-next.17
+
+## 5.0.0-next.16
+
+### Minor Changes
+
+- 76c8ded: Add the `symbols-no-detached` rule.
+
+### Patch Changes
+
+- Updated dependencies [8351103]
+  - @sketch-hq/sketch-assistant-types@5.0.0-next.16
+
+## 5.0.0-next.15
+
+### Major Changes
+
+- e76fabd: Introduction of the Assistants CLI.
+
+### Patch Changes
+
+- Updated dependencies [e76fabd]
+  - @sketch-hq/sketch-assistant-types@5.0.0-next.15
+
 ## 5.0.0-next.14
 
 ## 5.0.0-next.13